Enable access to core functionality through interface
To facilitate testing of the frontend independently from the core, it
was necessary to mock the core functionalities. However, these
functionalities were provided by package-level functions, which are not
easily mockable. Therefore, I have introduced an interface to declare
these functionalities. Although it is unlikely that other
implementations besides the mock will be needed, this approach allows
for flexibility. Additionally, it enables the possibility of having
multiple cores with different storage types, as no more package-level
global variables are used.