Move tests to the `_test` packages
Tests are moved to the `_test` package for improved encapsulation. Private
functions within the target package are not exposed during testing.
Additionally, mock files are suffixed with `_test` to be automatically
ignored by coverage tools. The exception to this rule is tests related
to the database. These tests remain in the same package as the database
functions to allow access to the internal database handle for checking
the state of the database.